Refactor into modules and add CLI
