+++ persona_generator.py
@@ -5,11 +5,6 @@ persona_generator.py - A tool to generate AI personas and add them to personas.j
 This script generates creative and diverse AI personas using the same LLM models
 available in rich_chat.py (Ollama and LM Studio). Generated personas are
 automatically added to the personas.json file used by rich_chat.py.
-
-Usage:
-    python persona_generator.py
-
-Features:
 - Uses the same LLM providers as rich_chat.py
 - Customizable persona attributes
 - Safety filters to ensure appropriate content
@@ -23,6 +18,7 @@ import time
 import logging
 import requests
 from typing import Dict, List, Any, Optional
+from api_clients import APIClient, OllamaClient, LMStudioClient, OpenRouterClient, OpenAIClient
 
 from rich.console import Console
 from rich.panel import Panel
@@ -45,302 +41,6 @@ logging.basicConfig(
     datefmt="[%X]",
 )
 log = logging.getLogger("persona_generator")
-
-class APIClient:
-    """Base class for LLM API clients."""
-
-    def __init__(self, name: str):
-        self.name = name
-
-    def generate_response(self, prompt: str, system: str = "") -> str:
-        """Generate a response from the LLM API."""
-        raise NotImplementedError("Subclasses must implement this method")
-
-    def get_available_models(self) -> List[str]:
-        """Get list of available models from this provider."""
-        raise NotImplementedError("Subclasses must implement this method")
-
-
-class OllamaClient(APIClient):
-    """Client for Ollama API."""
-
-    def __init__(self, base_url: str = "http://127.0.0.1:11434"):
-        super().__init__("Ollama")
-        self.base_url = base_url
-        self.api_url = f"{base_url}/api"
-        self.model = None
-
-    def set_model(self, model_name: str):
-        """Set the model to use for generation."""
-        self.model = model_name
-
-    def generate_response(self, prompt: str, system: str = "") -> str:
-        """Generate a response from Ollama API."""
-        if not self.model:
-            raise ValueError("Model must be set before generating responses")
-
-        # Prepare the chat format expected by Ollama
-        messages = []
-
-        # Add system message if provided
-        if system:
-            messages.append({"role": "system", "content": system})
-
-        # Add the user prompt
-        messages.append({"role": "user", "content": prompt})
-
-        # Make API request
-        try:
-            response = requests.post(
-                f"{self.api_url}/chat",
-                json={
-                    "model": self.model,
-                    "messages": messages,
-                    "stream": False
-                },
-                timeout=60
-            )
-            response.raise_for_status()
-            result = response.json()
-            return result["message"]["content"]
-        except requests.RequestException as e:
-            log.error(f"Ollama API error: {str(e)}")
-            return f"[Error generating response: {str(e)}]"
-
-    def get_available_models(self) -> List[str]:
-        """Get list of available models from Ollama."""
-        try:
-            response = requests.get(f"{self.api_url}/tags", timeout=10)
-            response.raise_for_status()
-            models = response.json().get("models", [])
-            return [model["name"] for model in models]
-        except requests.RequestException as e:
-            log.error(f"Failed to get Ollama models: {str(e)}")
-            return []
-
-
-class LMStudioClient(APIClient):
-    """Client for LM Studio API."""
-
-    def __init__(self, base_url: str = "http://127.0.0.1:1234/v1"):
-        super().__init__("LM Studio")
-        self.base_url = base_url
-        self.model = None
-
-    def set_model(self, model_name: str):
-        """Set the model to use for generation."""
-        self.model = model_name
-
-    def generate_response(self, prompt: str, system: str = "") -> str:
-        """Generate a response from LM Studio API."""
-        if not self.model:
-            raise ValueError("Model must be set before generating responses")
-
-        # Prepare the conversation format (OpenAI-compatible)
-        messages = []
-
-        # Add system message if provided
-        if system:
-            messages.append({"role": "system", "content": system})
-
-        # Add the user prompt
-        messages.append({"role": "user", "content": prompt})
-
-        # Make API request
-        try:
-            response = requests.post(
-                f"{self.base_url}/chat/completions",
-                json={
-                    "model": self.model,
-                    "messages": messages,
-                    "stream": False
-                },
-                timeout=60
-            )
-            response.raise_for_status()
-            result = response.json()
-            return result["choices"][0]["message"]["content"]
-        except requests.RequestException as e:
-            log.error(f"LM Studio API error: {str(e)}")
-            return f"[Error generating response: {str(e)}]"
-
-    def get_available_models(self) -> List[str]:
-        """Get list of available models from LM Studio."""
-        try:
-            response = requests.get(f"{self.base_url}/models", timeout=10)
-            response.raise_for_status()
-            models = response.json().get("data", [])
-            return [model["id"] for model in models]
-        except requests.RequestException as e:
-            log.error(f"Failed to get LM Studio models: {str(e)}")
-            return []
-
-
-class OpenRouterClient(APIClient):
-    """Client for OpenRouter API."""
-    
-    def __init__(self, api_key=""):
-        super().__init__("OpenRouter")
-        self.base_url = "https://openrouter.ai/api/v1"
-        self.api_key = api_key
-        self.model = None
-        self.update_headers()
-
-    def update_headers(self):
-        """Update headers with current API key."""
-        self.headers = {
-            'Authorization': f'Bearer {self.api_key}',
-            'Content-Type': 'application/json'
-        }
-
-    def set_model(self, model_name: str):
-        """Set the model to use for generation."""
-        self.model = model_name
-
-    def generate_response(self, prompt: str, system: str = "") -> str:
-        """Generate a response from OpenRouter API."""
-        if not self.api_key:
-            raise ValueError("OpenRouter API key not set")
-        if not self.model:
-            raise ValueError("Model must be set before generating responses")
-
-        try:
-            messages = [
-                {"role": "system", "content": system}
-            ]
-            
-            # Add the user prompt
-            messages.append({"role": "user", "content": prompt})
-
-            data = {
-                "model": self.model,
-                "messages": messages,
-                "temperature": 0.7,
-                "max_tokens": 2000
-            }
-            
-            response = requests.post(
-                f"{self.base_url}/chat/completions",
-                headers=self.headers,
-                json=data,
-                timeout=30
-            )
-            
-            response.raise_for_status()
-            
-            result = response.json()
-            
-            return result['choices'][0]['message']['content'].strip()
-        except requests.exceptions.RequestException as e:
-            log.error(f"[OpenRouter] Request failed: {e}")
-            return f"[Error generating response: {str(e)}]"
-        except Exception as e:
-            log.error(f"[OpenRouter] Error processing response: {str(e)}")
-            return f"[Error generating response: {str(e)}]"
-
-    def get_available_models(self) -> List[str]:
-        """Get list of available models from OpenRouter."""
-        if not self.api_key:
-            log.error("OpenRouter API key not set")
-            return []
-
-        try:
-            response = requests.get(
-                f"{self.base_url}/models",
-                headers=self.headers,
-                timeout=10
-            )
-            response.raise_for_status()
-            data = response.json()
-            return [model['id'] for model in data.get('data', [])]
-        except Exception as e:
-            log.error(f"Error fetching OpenRouter models: {str(e)}")
-            return []
-
-
-class OpenAIClient(APIClient):
-    """Client for OpenAI API."""
-    
-    def __init__(self, api_key=""):
-        super().__init__("OpenAI")
-        self.base_url = "https://api.openai.com/v1"
-        self.api_key = api_key
-        self.model = None
-        self.update_headers()
-
-    def update_headers(self):
-        """Update headers with current API key."""
-        self.headers = {
-            'Authorization': f'Bearer {self.api_key}',
-            'Content-Type': 'application/json'
-        }
-
-    def set_model(self, model_name: str):
-        """Set the model to use for generation."""
-        self.model = model_name
-
-    def generate_response(self, prompt: str, system: str = "") -> str:
-        """Generate a response from OpenAI API."""
-        if not self.api_key:
-            raise ValueError("OpenAI API key not set")
-        if not self.model:
-            raise ValueError("Model must be set before generating responses")
-
-        try:
-            messages = [
-                {"role": "system", "content": system}
-            ]
-            
-            # Add the user prompt
-            messages.append({"role": "user", "content": prompt})
-
-            data = {
-                "model": self.model,
-                "messages": messages,
-                "temperature": 0.7,
-                "max_tokens": 2000
-            }
-            
-            response = requests.post(
-                f"{self.base_url}/chat/completions",
-                headers=self.headers,
-                json=data,
-                timeout=30
-            )
-            
-            response.raise_for_status()
-            
-            result = response.json()
-            
-            return result['choices'][0]['message']['content'].strip()
-        except requests.exceptions.RequestException as e:
-            log.error(f"[OpenAI] Request failed: {e}")
-            return f"[Error generating response: {str(e)}]"
-        except Exception as e:
-            log.error(f"[OpenAI] Error processing response: {str(e)}")
-            return f"[Error generating response: {str(e)}]"
-
-    def get_available_models(self) -> List[str]:
-        """Get list of available models from OpenAI."""
-        if not self.api_key:
-            log.error("OpenAI API key not set")
-            return []
-
-        try:
-            response = requests.get(
-                f"{self.base_url}/models",
-                headers=self.headers,
-                timeout=10
-            )
-            response.raise_for_status()
-            data = response.json()
-            # Filter for chat completion models
-            models = [model['id'] for model in data.get('data', []) if "gpt" in model['id']]
-            return sorted(models)
-        except Exception as e:
-            log.error(f"Error fetching OpenAI models: {str(e)}")
-            return []
-
 
 class PersonaGenerator:
     """Generates AI personas using LLM providers."""
